# client/public/swagger.yaml
openapi: 3.0.0
info:
  description: Api documentation
  version: 1.0.1
  title: aim-container API
...
servers:
  - url: http://127.0.0.1:43801/{basePath}
    variables:
      basePath:
        default: api/v1
    description: Local env
security:
  - tokenAuth:
paths:
...
  "/projects/{experiment_name}/{commit_id}":
    get:
      operationId: getExperiment
      summary: Get project experiment
      tags:
        - Project
      parameters:
        - name: experiment_name
          in: path
          required: true
          schema:
            type: string
        - name: commit_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExperiment"
  "/projects/insight/{insight_name}":
    get:
      operationId: getProjectInsight
      summary: Get project insight
      tags:
        - Project
      parameters:
        - name: insight_name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectInsight"
  "/projects/{experiment_name}/{commit_id}/{path}":
    get:
      operationId: getExperimentComponent
      summary: Get component content
      tags:
        - Project
      parameters:
        - name: experiment_name
          in: path
          required: true
          schema:
            type: string
        - name: commit_id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    ProjectInsight:
      type: object
    ProjectExperiment:
      type: object
...